Improve: IOCP on win, blackholes, remote initialization.
This is set of refactoring made trying to fix #348.

IOCP: now if we cancel a task on active socket, we steal r/w
 buffers from active connection and leave them intact for iocp
 queue. After the queue abandoned they will be released. The reason
 is that overlapped io is low-level operation, and in process we
 must not change (say, free) any invoked buffers until it finished.

Blackholes: more carefully check to not grab any object which may
be released from outside. Blackholes are just a 'fire' queries; we
just send them and never wait for an answer. Also we never retry on
fail. So, having just output buffer to perform sending must be
absolutely enough for them to work. And so, any other essences like
reply parser, logger, success reporter must be explicitly
removed/deleted to avoid any confusion.

Initialization: when a task fails we either finish the connection,
either schedule next retry on it. In any branch we need to unlink
task from connection (and then m.b. create new task). Finally the
whole connection will be released/removed.

+++ src/sphinx.cpp
@@ -1587,13 +1587,13 @@ static size_t sphReadThrottled ( int iFD, void * pBuf, size_t iCount )
 	if ( iCount<=0 )
 		return iCount;
 
-	auto iStep = g_iMaxIOSize ? Min ( iCount, g_iMaxIOSize ) : iCount;
+	auto iStep = g_iMaxIOSize ? Min ( iCount, (size_t)g_iMaxIOSize ) : iCount;
 	auto * p = ( BYTE * ) pBuf;
 	size_t nBytesToRead = iCount;
 	while ( iCount )
 	{
 		ThrottleSleep();
-		auto iChunk = Min ( iCount, iStep );
+		auto iChunk = (long) Min ( iCount, iStep );
 		auto iRead = sphRead ( iFD, p, iChunk );
 		p += iRead;
 		iCount -= iRead;
